filled in missing tests

+++ tests/test_generate_graph.cpp
@@ -111,7 +111,7 @@ TEST_CASE("Test generate_label_based_graph basic functionality")
     generate_label_based_graph(graph, coords, F);
 
     // check that all nodes were created
-    REQUIRE(graph.getNodeCount() == coords.size());
+    REQUIRE(graph.getNodeCount() == static_cast<int>(coords.size()));
 
     // check that each node's coordinates are correct
     for (size_t i = 0; i < coords.size(); ++i)
